Applied fuzzily described algorithm correctly, verified by our own implementation which follows the algorithm described in Engineering a Compiler SE, page 106, figure 3.8
When calculating the FOLLOW set for a certain non-terminal NT, you have to find all occurrences of that non-terminal in all RHS to find the symbols that could FOLLOW. For each occurrence found, you look up the symbol's FIRST set FOLLOWING after that occurring non-terminal, unless that is the last symbol of the RHS - than you add the FOLLOW set of the LHS. The FIRST set of that FOLLOWING symbol can contain an epsilon. In that case, also the FIRST set of the symbol following after the next symbol of that certain non-terminal NT should be added and if that contains an epsilon as well, this process should be repeated. If in any case, the last symbol is reached, the FOLLOW set of the LHS should be added, as usual.

Here, this occurs in two places:
* when calculating FOLLOW for F since:
- E  -> F E'
- E' -> epsilon
Given E' is the last symbol, we add the FOLLOW set of E to the FOLLOW set of F.

* when calculating FOLLOW for A since:
- G  -> A G'
- G' -> epsilon
Given G' is the last symbol, we add the FOLLOW set of G to the FOLLOW set of A.

In code, this is what happens: https://github.com/tcoenraad/compiler-construction/blob/master/block-2/pp/block2/cc/ll/AwesomeLLCalc.java#L112